+++ include/sensor_data/loop_pose.hpp
@@ -16,10 +16,10 @@ public:
   double time = 0.0;
   unsigned int index0 = 0;
   unsigned int index1 = 0;
-  Eigen::Matrix4f pose = Eigen::Matrix4f::Identity();
+  Eigen::Matrix4d pose = Eigen::Matrix4d::Identity();
 
 public:
-  Eigen::Quaternionf getQuaternion();
+  Eigen::Quaterniond getQuaternion();
 };
 } // namespace avp_mapping
